Update in UserMovieUseCases

+++ webApi/Application/UseCases/UserMovieUseCases.cs
@@ -10,11 +10,9 @@ namespace webApi.Application.UseCases
         private readonly IUserMovieRepository _repository = repository;
         public async Task<bool> AddMovieOnUser(UserMovieDto userMovieDto)
         {
-            var movieDB = await _repository.CheckExistMovie(userMovieDto.MovieId);
-            if (movieDB == null) return false;
-
-            var userDB = await _repository.CheckExistUser(userMovieDto.UserId);
-            if (userDB == null) return false;
+            var movieDB = await CheckExistingMovie(userMovieDto.MovieId);
+            var userDB = await CheckExistingUser(userMovieDto.UserId);
+            if (userDB == null || movieDB == null) return false;
 
             _repository.AddMovieOnUser(userDB, movieDB);
             return await _repository.SaveChangeAsync();
@@ -27,14 +25,22 @@ namespace webApi.Application.UseCases
 
         public async Task<bool> RemoveMovieOnUser(UserMovieDto userMovieDto)
         {
-            var movieDB = await _repository.CheckExistMovie(userMovieDto.MovieId);
-            if (movieDB == null) return false;
-
-            var userDB = await _repository.CheckExistUser(userMovieDto.UserId);
-            if (userDB == null) return false;
+            var movieDB = await CheckExistingMovie(userMovieDto.MovieId);
+            var userDB = await CheckExistingUser(userMovieDto.UserId);
+            if (userDB == null || movieDB == null) return false;
 
             _repository.RemoveMovieOnUser(userDB, movieDB);
             return await _repository.SaveChangeAsync();
+        }
+
+        private async Task<Movie?> CheckExistingMovie(Guid movieId)
+        {
+            return await _repository.CheckExistMovie(movieId);
+        }
+
+        private async Task<User?> CheckExistingUser(Guid userId)
+        {
+            return await _repository.CheckExistUser(userId);
         }
     }
 }
